Some socket refactoring
