+++ client/src/components/game/Chat.js
@@ -1,8 +1,6 @@
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import Box from "@material-ui/core/Box";
-import io from "socket.io-client";
-const socket = io("localhost:5000");
 
 const useStyles = makeStyles((theme) => ({
   chat: {
