+++ client/src/services/socketService.js
@@ -1,0 +1,15 @@
+import io from "socket.io-client";
+const socket = io("localhost:5000");
+
+export const sendMessage = (user, message) => {
+  socket.emit(
+    "enviarMensaje",
+    {
+      usuario: user,
+      mensaje: message,
+    },
+    function (resp) {
+      console.log("respuesta server: ", resp);
+    }
+  );
+};
